Move the heatmaps overrides
Easier than fighting all the rollup issues with different leaflet instances
