+++ app/assets/config/manifest.js
@@ -6,3 +6,4 @@
 //= link_tree ../builds
 //= link_tree ../../../node_modules/openseadragon/build/openseadragon/images
 //= link blacklight_gallery/manifest.js
+//= link L.Control.Sidebar.js
